requirements.txt
boto3
mail-parser
pydantic
python-dotenv
supabase
requests
fastapi
uvicorn
cryptography
playwright
google-api-python-client
google-auth-oauthlib
google-auth

asyncpg
sqlalchemy[asyncio]
alembic
pydantic-settings